pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Bug Maker"]

[tool.poetry.dependencies]
python = "3.8.10"


[tool.poetry.dev-dependencies]
python-telegram-bot = "13.11"
pyyaml = "5.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
